Tests for object initializer and private forward.
Some code bases use "object initalizer" classes. A common approach
is to document the class in literal notation.
See
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Object_initializer

Also - some code bases use public classes which have public methods
that forward to a private class. The private class generally has the
jsdoc annotations. In this case, the documentation should use the
public class as the name, but take the documentation on the methods
in the private class.