Use tempfile for subprocess stdout to handle large output volumes
When a subprocess command generates a lot of output to stdout (exceeds
the OS pipe buffer), using checkout_output will end up returning an
incomplete stdout (only what fits in the buffer).

Similar issues exist with Popen.communicate(), see:
https://docs.python.org/3/library/subprocess.html#subprocess.Popen.communicate

The simplest solution is to use Popen instead of check_output and send
stdout to a tempfile. Note: this will possibly also block for large
stderr output, although that is unexpected for a jsdoc call. See:
https://stackoverflow.com/questions/3575554
https://stackoverflow.com/questions/1180606