Support freeform text @see tag
Adds @see tags with freeform text values, as specified in the
JSDoc manual: http://usejsdoc.org/tags-see.html

The extant support seems to interpret values as references to
Sphinx entities rather than JSDoc namepaths. This change keeps
that behavior for references.

Interprets the value as text if it contains any space characters.
This is a bit of a hack but should work for most cases. If JSDoc
namepath support is ever added, it could be changed to attempt
namepath resolution and failover to text for consistency with
JSDoc behavior.