Remove the rest of the `rst.escape()` calls.
Tests pass more than before.

We shouldn't need to escape in `or_types()`, since `|` is not special to RST if it's not the first thing on a line and no valid TS or JS chars that show up in types should throw off the RST parser either.

Maybe there are weird corner cases where we'll need to add a real escaper, but we never had one before, so it shouldn't be a regression.