Put an explicit `constructor` field on Class.
Otherwise, we'd end up conflating fields on Class, like `example`, with the same ones on the default constructor. Some of them are semantically distinct, and jsdoc (and typedoc, I believe) extracts them distinctly. Plus we'd have to either repeat fields from Function or subclass Function. The latter would be fine, but the conflation concerns remain.