Remove the uninformative fields from the jsdoc class test.
The idea is to make the tests less brittle and more relevant to read.

I tried doing it by adding default values to the IR, but dataclasses get grumpy if defaultless fields follow default-having ones, and there's no pretty way to get them to use keyword-only args for everything. So we just use a million asserts.