Think out the sourcing and display philosophy for params.
Because we now analyze up front, we choose only once where to harvest params from. We can't grab only documented params at once place in the renderer and then descend into the param-name-only metadata at another. So we implement a clear hierarchy of sources for the IR and filter out some uninformative things later in the renderer.

Specifically...

* We no longer show args in the arg list that are utterly uninformative, lacking both description and type info.
* Consequently, add "Snorf" so a_ will show up in the output and prove that it isn't an RST injection.