Replace make_longname with make_path_segments, and get some of it under test. It's about 60% done.
I'm trying to make it more obviously correct. I'm also trying to go straight from the TS JSON to path segments, without building up a string and putting it through the parser as the JSDoc analyzer does. The more nested TS JSON, which has no already concatenated pathnames, is more amenable to this.

* Fix some mistaken documentation: the path segment delimiter between files and top-level constructs should be ".", not "/".
* Remove some extraneous properties from the IndexByIdTests fixture.
* Comment out some parts of index_by_id that seem to do nothing. I'll be surer once convert_nodes() is under test; then I can delete them properly.