Change ``.types`` in the IR to ``.type``. There's no reason for unions to be special.
Getting ready to make sure even fancy TS types can be rendered as text, it occurred to me that union types are not special and should not be elevated to a data-structure-dominating thing in the IR. Thus, change the arrays that represented types in the IR to just strings. The lists were of little practical use as well. The only place that used them was where we guessed the type of an ambiguously typed default in the JSDoc analyzer. That was easily remedied.