Rewrite large parts of the suffix tree.
Fix several crashers. Most of the problems stemmed from not realizing (and correctly dealing with or reporting) the possibility of a node holding both a value and a further subtree. The new implementation is based on a sounder understanding of the state space and is better tested and explained.

I'm hoping this also fixes https://github.com/mozilla/sphinx-js/pull/169#issuecomment-779092066.