update Sphinx to 4.x
sphinx relies on other deprecated Jinja functions that are now removed, such that in order to get a working build we have to update Sphinx to 4.x at the same time.

Note that this introduced behavioral changes that break our tests: `"type()"` is returned instead of `*type*` and `autofunction_static` no longer works. This at least installs deps and gets to the testrunner before it falls over, which is an improvement