Test incremental builds.
This test leverages the existing Sphinx `pytest` infrastructure, but
doing so requires a different file layout for the test inputs.  It's
probably possible to follow the existing layout but it's not worth
much effort.  Note that since the input files are modified, they can't
be used in-place in the manner of the existing tests.