Fix make_xref so literals don't get rendered as functions (#209)
Sphinx 4.1.0 changed sphinx.utils.docfields Field.make_xref such
that when sphinx-js documents a Javascript function with a parameter,
the type gets rendered as a function. This affects TypedField
(arguments) and GroupedField (throws).

This patches the Sphinx javascript domain so that it uses a JSXrefMixin
to fix make_xref just like PythonXrefMixin does and then defines
JStypedField and JSGroupedField classes using it. This sets inliner to
None before calling Field.make_xref which prevents argument types from
being rendered as functions.

This was tested with Sphinx 4.1.0 (which introduced the inliner stuff)
through 5.3.0 (current version as of this writing).

This also fixes requirements_dev.txt and setup.py install_requires so
that dependencies are defined in setup.py where possible and not
duplicated in multiple places making it hard to pin down which thing is
getting tested.