Drop Sphinx upper-bound requirement (#227)
This drops the upper bound for the Sphinx requirement because I'm pretty
sure we don't need it. We can fix issues as they come up and we aren't
helped by having an upper bound.

This also reworks the matrix for tox.ini so that it tests with multiple
versions of Sphinx.